requirements.txt
certifi==2025.1.31
httpx==0.24.1
python-dotenv>=0.19.0
python-telegram-bot>=20.0
nest-asyncio==1.5.6
imgkit==1.2.3
pytz==2023.3
schedule==1.2.2
openai==1.52.2
requests==2.32.3
pyshorteners==1.0.1
pyTelegramBotAPI==4.26.0
yt-dlp>=2023.0.0
browser_cookie3==0.20.1
aiohttp>=3.8.0
flake8==7.1.2
pytest==7.0.0
pycodestyle==2.12.1
pyflakes==3.2.0
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.2
beautifulsoup4==4.12.2
python-dateutil==2.9.0
# Development tools
autopep8==2.3.2
black==25.1.0


# API-specific
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0